@@ -16,7 +16,8 @@
 package com.popmovies.utilities;
 
 import android.net.Uri;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import com.popmovies.Config;
 
